+++ lib/src/elements/sink/mod.rs
@@ -60,6 +60,11 @@ pub use renderer::{
 pub use renderer::{
     PipeWireAudioRenderer, PipeWireAudioRendererError, PipeWireAudioRendererOptions,
 };
+#[cfg(any(
+    all(target_os = "windows", any(feature = "d3d11", feature = "d3d12")),
+    all(target_os = "linux", feature = "vulkan")
+))]
+pub use renderer::{VideoWindow, VideoWindowError};
 #[cfg(all(target_os = "linux", feature = "vulkan"))]
 pub use renderer::{VulkanWindowRenderer, VulkanWindowRendererError, WindowSize};
 #[cfg(all(target_os = "windows", feature = "wasapi-renderer"))]