+++ lib/src/elements/sink/renderer/mod.rs
@@ -10,6 +10,11 @@ mod cuda;
 #[cfg(all(target_os = "linux", feature = "pipewire-audio-renderer"))]
 mod linux;
 mod submit_error;
+#[cfg(any(
+    all(target_os = "windows", any(feature = "d3d11", feature = "d3d12")),
+    all(target_os = "linux", feature = "vulkan")
+))]
+mod video_window;
 #[cfg(all(target_os = "linux", feature = "vulkan"))]
 mod vulkan;
 #[cfg(any(
@@ -28,6 +33,11 @@ pub use cuda::*;
 #[cfg(all(target_os = "linux", feature = "pipewire-audio-renderer"))]
 pub use linux::*;
 pub use submit_error::SubmitError;
+#[cfg(any(
+    all(target_os = "windows", any(feature = "d3d11", feature = "d3d12")),
+    all(target_os = "linux", feature = "vulkan")
+))]
+pub use video_window::{VideoWindow, VideoWindowError};
 #[cfg(all(target_os = "linux", feature = "vulkan"))]
 pub use vulkan::*;
 #[cfg(any(